extract sourceTimeSeries.js
