+++ gatsby-node.js
@@ -12,7 +12,6 @@ const {
   getGrassroots,
   // getFundraisers,
   getResources,
-  loadTimeSeries,
 } = require(`./scripts/fetchData.js`);
 const { validStatePages, slugify } = require(`./scripts/utils.js`);
 
@@ -60,10 +59,6 @@ exports.sourceNodes = async (params) => {
     ["Resources", getResources],
   ];
 
-  console.log("Load time series");
-  // NOTE: loads asynchronously
-  loadTimeSeries();
-  
   for (let i = 0; i < nodes.length; i++) {
     await createSourceNodes(nodes[i][0], nodes[i][1], params);
   }
